+++ CHANGELOG.md
@@ -2,28 +2,39 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).c
+
+## [1.4.0] - 2020-03-18
+### Added
+
+- Now Altimetry object contains also the elevation related kilometer
 
 ## [1.3.1] - 2020-03-13
 ### Fixed
+
 - Updated README.md to show hot to use new implemented method on v1.3.0
 
 ## [1.3.0] - 2020-03-13
 ### Added
+
 - GetGPXAltimetry() method to retrieve a complete altimetry of the trace
 
 ## [1.2.0] - 2019-02-01
 ### Added
+
 - Method that returns a list of latitude and longitudes
 
 ### Updated
+
 - README.md
 
 ## [1.1.0] - 2019-12-30
 ### Added
+
 - Distance calculation in kilometers
 - Code snippet to use the library
 
 ## [1.0.0] - 2019-12-28
 ### Added
+
 - First implemetation